+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
+## Unreleased
+
+### Fixed
+- Disabled current estimation no longer publishes a zero-current value during plugin startup.
+
 ## [2.3.4] - 2026-09-17
 
 ### Fixed